Changed from int to uint8_t to ensure positive values and smaller size

+++ src/creatures/creature.cpp
@@ -1,8 +1,9 @@
 #include "creatures/creature.h"
 #include <iostream>
 
-Creature::Creature(std::string name, int hp, int attack, int defense, int speed)
-    : name(name), hp(hp), maxHp(hp), attack(attack),
+Creature::Creature(
+    std::string name, uint8_t hp, uint8_t attack, uint8_t defense, uint8_t speed
+) : name(name), hp(hp), maxHp(hp), attack(attack),
     defense(defense), speed(speed), moveCount(0),
     status(NONE), statusDuration(0) {}
 
@@ -10,8 +11,8 @@ bool Creature::operator==(const Creature &other) const {
     return (name == other.name);
 }
 
-void Creature::takeDamage(int dmg) {
-    int actual = (int)(((rand() % 5) + dmg) * (100 - defense) / 100);
+void Creature::takeDamage(uint8_t dmg) {
+    uint8_t actual = (uint8_t)(((rand() % 5) + dmg) * (100 - defense) / 100);
     if (actual < 1) {
         actual = 1;
     }
@@ -30,7 +31,7 @@ void Creature::displayStatus() const {
 }
 
 void Creature::displayMoves() const {
-    for (int i = 0; i < moveCount; i++) {
+    for (uint8_t i = 0; i < moveCount; i++) {
         std::cout << i + 1 << ". ";
         moves[i].display();
     }
@@ -43,7 +44,7 @@ void Creature::addMove(Move move) {
     }
 }
 
-void Creature::setStatus(Status s, int duration) {
+void Creature::setStatus(Status s, uint8_t duration) {
     if (status != NONE) {
         return;
     }
@@ -69,7 +70,7 @@ void Creature::applyStatusEffect() {
 
     switch (status) {
         case INFECTED: {
-            int dmg = maxHp * 0.10;
+            uint8_t dmg = maxHp * 0.10;
             hp -= dmg;
             if (hp < 0) {
                 hp = 0;
@@ -82,7 +83,7 @@ void Creature::applyStatusEffect() {
             break; 
         }
         case BURNED: {
-            int bdmg = maxHp * 0.05;
+            uint8_t bdmg = maxHp * 0.05;
             hp -= bdmg;
             if (hp < 0) {
                 hp = 0;
@@ -106,15 +107,15 @@ std::string Creature::getName() const {
     return name;
 }
 
-int Creature::getSpeed() const {
+uint8_t Creature::getSpeed() const {
     return speed;
 }
 
-int Creature::getAttack() const {
+uint8_t Creature::getAttack() const {
     return attack;
 }
 
-int Creature::getDefense() const {
+uint8_t Creature::getDefense() const {
     return defense;
 }
